Add undo functionality, validation for edit mode, and support for decimal shift hours

+++ src/ShiftScheduler.API/Models/Schedule.cs
@@ -39,5 +39,5 @@ public class DaySchedule
 /// </summary>
 public class SchedulerSettings
 {
-    public int ShiftHours { get; set; } = 8;
+    public double ShiftHours { get; set; } = 8;
 }
